move cache and adapter into delayln component
